python: refactor run_test, compile_prereqs so only python stdlib used
declarative: tests/url.ini defines test downloads
