+++ cmake/test_setup.cmake
@@ -58,7 +58,7 @@ set(_nml ${CMAKE_CURRENT_SOURCE_DIR}/tests/data/test${TESTNAME}/inputs/config.nm
 #       WORKING_DIRECTORY ${PROJECT_SOURCE_DIR})
 # else()
 add_test(NAME gemini:${TESTNAME}
-  COMMAND ${Python3_EXECUTABLE} ${CMAKE_CURRENT_SOURCE_DIR}/scripts/meson_run_test.py ${TESTNAME} ${MPIEXEC_EXECUTABLE} $<TARGET_FILE:gemini.bin> ${_nml} ${_outdir})
+  COMMAND ${Python3_EXECUTABLE} ${CMAKE_CURRENT_SOURCE_DIR}/scripts/run_test.py ${TESTNAME} ${MPIEXEC_EXECUTABLE} $<TARGET_FILE:gemini.bin> ${_nml} ${_outdir})
 
 
 # NOTE: don't use REQUIRED_FILES because it won't let file download if not present.
