+++ gemini3d/__init__.py
@@ -1,5 +1,5 @@
 from .readdata import read_config, readgrid, readdata, read_Efield
 from .output_compare import compare_all
 from .raw import read4D, read3D, read2D, read_time
-from .utils import get_cpu_count, get_mpi_count, url_retrieve, extract_zip, extract_tar
+from .utils import get_cpu_count, get_mpi_count
 from .base import get_simsize
