+++ scripts/compile_prereqs_gcc.py
@@ -6,21 +6,33 @@ import subprocess
 import shutil
 import logging
 from pathlib import Path
-import pkg_resources
 from argparse import ArgumentParser
-import typing
+import typing as T
 import sys
 import os
 from functools import lru_cache
-import gemini3d
+
+from web import url_retrieve, extract_tar
+
+try:
+    import pkg_resources
+except ModuleNotFoundError:
+    pkg_resources = None
+try:
+    from gemini3d.utils import get_cpu_count
+except ImportError:
+
+    def get_cpu_count() -> int:
+        return 1
+
 
 # ========= user parameters ======================
 BUILDDIR = "build"
-NJOBS = gemini3d.get_cpu_count()
+NJOBS = get_cpu_count()
 # Library parameters
 HDF5VERSION = "1.10.6"
-HDF5URL = "https://zenodo.org/record/3659270/files/hdf5-1.10.6.tar.bz2?download=1"
-HDF5MD5 = "03095102a6118c32a75a9b9b40be66f2"
+HDF5URL = "https://zenodo.org/record/3700903/files/hdf5-1.12.0.tar.bz2?download=1"
+HDF5MD5 = "1fa68c4b11b6ef7a9d72ffa55995f898"
 HDF5DIR = f"hdf5-{HDF5VERSION}"
 
 MPIVERSION = "3.1.5"  # OpenMPI 4 needs Scalapack 2.1
@@ -41,7 +53,7 @@ MUMPSDIR = "mumps"
 nice = ["nice"] if sys.platform == "linux" else []
 
 
-def hdf5(dirs: typing.Dict[str, Path]):
+def hdf5(dirs: T.Dict[str, Path], env: T.Mapping[str, str] = None):
     """ build and install HDF5 """
     if os.name == "nt":
         raise SystemExit("Please use binaries from HDF Group for Windows appropriate for your compiler.")
@@ -50,33 +62,35 @@ def hdf5(dirs: typing.Dict[str, Path]):
     source_dir = dirs["workdir"] / HDF5DIR
 
     tarfn = f"hdf5-{HDF5VERSION}.tar.bz2"
-    gemini3d.url_retrieve(HDF5URL, tarfn, ("md5", HDF5MD5))
-    gemini3d.extract_tar(tarfn, source_dir)
+    url_retrieve(HDF5URL, tarfn, ("md5", HDF5MD5))
+    extract_tar(tarfn, source_dir)
 
-    env = get_compilers()
+    if not env:
+        env = get_compilers()
 
     cmd = nice + ["./configure", f"--prefix={install_dir}", "--enable-fortran", "--enable-build-mode=production"]
 
     subprocess.check_call(cmd, cwd=source_dir, env=env)
 
-    cmd = nice + ["make", "-C", str(source_dir), f"-j{NJOBS}", "install"]
+    cmd = nice + ["make", "-C", str(source_dir), f"-j {NJOBS}", "install"]
     subprocess.check_call(cmd)
 
 
-def openmpi(dirs: typing.Dict[str, Path]):
+def openmpi(dirs: T.Dict[str, Path], env: T.Mapping[str, str] = None):
     """ build and install OpenMPI """
     if os.name == "nt":
-        raise SystemExit("OpenMPI is not available in native Windows.")
+        raise SystemExit("OpenMPI is not available in native Windows. Use MS-MPI instead.")
 
     install_dir = dirs["prefix"] / MPIDIR
     source_dir = dirs["workdir"] / MPIDIR
 
     tarfn = f"openmpi-{MPIVERSION}.tar.bz2"
     url = f"https://download.open-mpi.org/release/open-mpi/v{MPIVERSION[:3]}/{tarfn}"
-    gemini3d.url_retrieve(url, tarfn, ("sha1", MPISHA1))
-    gemini3d.extract_tar(tarfn, source_dir)
+    url_retrieve(url, tarfn, ("sha1", MPISHA1))
+    extract_tar(tarfn, source_dir)
 
-    env = get_compilers()
+    if not env:
+        env = get_compilers()
 
     cmd = nice + ["./configure", f"--prefix={install_dir}", f"CC={env['CC']}", f"CXX={env['CXX']}", f"FC={env['FC']}"]
 
@@ -86,7 +100,7 @@ def openmpi(dirs: typing.Dict[str, Path]):
     subprocess.check_call(cmd)
 
 
-def lapack(wipe: bool, dirs: typing.Dict[str, Path], buildsys: str):
+def lapack(wipe: bool, dirs: T.Dict[str, Path], buildsys: str):
     """ build and insall Lapack """
     install_dir = dirs["prefix"] / LAPACKDIR
     source_dir = dirs["workdir"] / LAPACKDIR
@@ -104,7 +118,7 @@ def lapack(wipe: bool, dirs: typing.Dict[str, Path], buildsys: str):
         raise ValueError(f"unknown build system {buildsys}")
 
 
-def scalapack(wipe: bool, dirs: typing.Dict[str, Path], buildsys: str):
+def scalapack(wipe: bool, dirs: T.Dict[str, Path], buildsys: str):
     """ build and install Scalapack """
     source_dir = dirs["workdir"] / SCALAPACKDIR
     build_dir = source_dir / BUILDDIR
@@ -123,7 +137,7 @@ def scalapack(wipe: bool, dirs: typing.Dict[str, Path], buildsys: str):
         raise ValueError(f"unknown build system {buildsys}")
 
 
-def mumps(wipe: bool, dirs: typing.Dict[str, Path], buildsys: str):
+def mumps(wipe: bool, dirs: T.Dict[str, Path], buildsys: str, env: T.Mapping[str, str] = None):
     """ build and install Mumps """
     install_dir = dirs["prefix"] / MUMPSDIR
     source_dir = dirs["workdir"] / MUMPSDIR
@@ -134,19 +148,23 @@ def mumps(wipe: bool, dirs: typing.Dict[str, Path], buildsys: str):
 
     update(source_dir, MUMPSGIT)
 
-    lib_args = [f"-DSCALAPACK_ROOT={scalapack_lib}", f"-DLAPACK_ROOT={lapack_lib}"]
+    if env and env["FC"] == "ifort":
+        lib_args = []
+    else:
+        env = get_compilers()
+        lib_args = [f"-DSCALAPACK_ROOT={scalapack_lib}", f"-DLAPACK_ROOT={lapack_lib}"]
 
     if buildsys == "cmake":
         args = [f"-DCMAKE_INSTALL_PREFIX={install_dir}"]
-        cmake_build(args + lib_args, source_dir, build_dir, wipe, env=get_compilers())
+        cmake_build(args + lib_args, source_dir, build_dir, wipe, env=env)
     elif buildsys == "meson":
         args = [f"--prefix={dirs['prefix']}"]
-        meson_build(args + lib_args, source_dir, build_dir, wipe, env=get_compilers())
+        meson_build(args + lib_args, source_dir, build_dir, wipe, env=env)
     else:
         raise ValueError(f"unknown build system {buildsys}")
 
 
-def cmake_build(args: typing.List[str], source_dir: Path, build_dir: Path, wipe: bool, env: typing.Mapping[str, str]):
+def cmake_build(args: T.List[str], source_dir: Path, build_dir: Path, wipe: bool, env: T.Mapping[str, str]):
     """ build and install with CMake """
     cmake = cmake_minimum_version("3.13")
     cachefile = build_dir / "CMakeCache.txt"
@@ -162,7 +180,7 @@ def cmake_build(args: typing.List[str], source_dir: Path, build_dir: Path, wipe:
     raise SystemExit(ret.returncode)
 
 
-def meson_build(args: typing.List[str], source_dir: Path, build_dir: Path, wipe: bool, env: typing.Mapping[str, str]):
+def meson_build(args: T.List[str], source_dir: Path, build_dir: Path, wipe: bool, env: T.Mapping[str, str]):
     """ build and install with Meson """
     meson = shutil.which("meson")
     if not meson:
@@ -191,6 +209,10 @@ def cmake_minimum_version(min_version: str = None) -> str:
 
     if not min_version:
         return cmake
+
+    if pkg_resources is None:
+        return cmake
+
     cmake_ver = subprocess.check_output([cmake, "--version"], universal_newlines=True).split("\n")[0].split(" ")[2]
     if pkg_resources.parse_version(cmake_ver) < pkg_resources.parse_version(min_version):
         logging.error(f"CMake {cmake_ver} is less than minimum required {min_version}")
@@ -217,8 +239,8 @@ def update(path: Path, repo: str):
 
 
 @lru_cache()
-def get_compilers() -> typing.Mapping[str, str]:
-
+def get_compilers() -> T.Mapping[str, str]:
+    """ get paths to GCC compilers """
     env = os.environ
 
     fc_name = "gfortran"
@@ -229,19 +251,19 @@ def get_compilers() -> typing.Mapping[str, str]:
     if fc_name not in fc:
         fc = shutil.which(fc_name)
     if not fc:
-        raise FileNotFoundError(f"{fc_name} not found")
+        raise FileNotFoundError(fc_name)
 
     cc = env.get("CC", "")
     if cc_name not in cc:
         cc = shutil.which(cc_name)
     if not cc:
-        raise FileNotFoundError(f"{cc_name} not found")
+        raise FileNotFoundError(cc_name)
 
     cxx = env.get("CXX", "")
     if cxx_name not in cxx:
         cxx = shutil.which(cxx_name)
     if not cxx:
-        raise FileNotFoundError(f"{cxx_name} not found")
+        raise FileNotFoundError(cxx_name)
 
     env.update({"FC": fc, "CC": cc, "CXX": cxx})
