+++ tests/url.ini
@@ -1,0 +1,15 @@
+[2d_fang]
+url = https://zenodo.org/record/3697235/files/test2d_fang.zip?download=1
+md5 = 7b237883b17d6da331b7db0309cc3da4
+
+[2d_glow]
+url = https://zenodo.org/record/3697235/files/test2d_glow.zip?download=1
+md5 = 4b8e31e239918ff9e02cbc3033099d60
+
+[3d_fang]
+url = https://zenodo.org/record/3698137/files/test3d_fang.zip?download=1
+md5 = a270586fbd41ff58e3a5ed0acac63708
+
+[3d_glow]
+url = https://zenodo.org/record/3698137/files/test3d_glow.zip?download=1
+md5 = 993d817ee847314a631396608df56e0c
